implements restrictions on non-whitelisted players

+++ code/modules/mob/new_player/new_player.dm
@@ -37,6 +37,8 @@
 
 
 /mob/new_player/proc/check_playerpolls()
+	if(!WHITELIST_CHECK(client))
+		return FALSE
 	var/output
 	if (SSdbcore.Connect())
 		var/isadmin = FALSE
@@ -74,6 +76,8 @@
 	. = ..()
 
 	if(!SSticker)
+		return
+	if(!WHITELIST_CHECK(client))
 		return
 
 	if(SSticker.current_state == GAME_STATE_PREGAME)
@@ -143,6 +147,9 @@
 			view_som()
 
 		if("SelectedJob")
+			if(!WHITELIST_CHECK(client))
+				WHITELIST_MESSAGE(client)
+				return
 			if(!SSticker)
 				return
 			if(!GLOB.enter_allowed)
@@ -207,6 +214,9 @@
 
 
 /mob/new_player/proc/late_choices()
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	var/list/dat = list("<div class='notice'>Round Duration: [DisplayTimeText(world.time - SSticker.round_start_time)]</div>")
 	if(!GLOB.enter_allowed)
 		dat += "<div class='notice red'>You may no longer join the round.</div><br>"
@@ -374,6 +384,9 @@
 
 /mob/new_player/proc/transfer_character()
 	. = new_character
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	if(.)
 		mind.transfer_to(new_character, TRUE) //Manually transfer the key to log them in
 		qdel(src)
@@ -399,6 +412,9 @@
 	return TRUE
 
 /mob/new_player/proc/try_to_observe()
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	if(!SSticker || SSticker.current_state == GAME_STATE_STARTUP)
 		to_chat(src, span_warning("The game is still setting up, please try again later."))
 		return
@@ -474,6 +490,10 @@
 		return
 	ready = !ready
 	if(ready)
+		if(!WHITELIST_CHECK(client))
+			WHITELIST_MESSAGE(client)
+			ready = FALSE
+			return
 		GLOB.ready_players += src
 	else
 		GLOB.ready_players -= src
@@ -481,6 +501,9 @@
 
 ///Attempts to latejoin the player
 /mob/new_player/proc/attempt_late_join(queue_override = FALSE)
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
+		return
 	if(!SSticker?.mode || SSticker.current_state != GAME_STATE_PLAYING)
 		to_chat(src, span_warning("The round is either not ready, or has already finished."))
 		return
@@ -516,6 +539,8 @@
 
 
 /mob/new_player/proc/take_ssd_mob()
+	if(!WHITELIST_CHECK(client))
+		WHITELIST_MESSAGE(client)
 	if((src.key in GLOB.key_to_time_of_death) && (GLOB.key_to_time_of_death[src.key] + TIME_BEFORE_TAKING_BODY > world.time))
 		to_chat(src, span_warning("You died too recently to be able to take a new mob."))
 		return
